+++ src/client/content/js/components/authenticated/activity/registered/list-item.js
@@ -1,5 +1,6 @@
 import React from 'react/addons';
 import {Navigation} from 'react-router';
+import moment from 'moment';
 
 export default React.createClass({
     mixins: [
@@ -19,6 +20,10 @@ export default React.createClass({
             <tr>
                 <td>{number}</td>
                 <td>{this.props.item.name}</td>
+                <td>{this.props.item.proximity}</td>
+                <td>{this.props.item.accuracy.toFixed(4)}</td>
+                <td>{this.props.item.measuredPower}</td>
+                <td>{moment(this.props.item.lastSeen).fromNow()}</td>
             </tr>
         );
     }