+++ src/client/content/js/models/activity.js
@@ -1,0 +1,12 @@
+import {Record} from 'immutable';
+
+export default Record({
+    name: '',
+    uuid: '',
+    major: 0,
+    minor: 0,
+    proximity: 0,
+    accuracy: 0,
+    measuredPower: 0,
+    lastSeen: null
+});